Move data table versioning utils into src/workspace-data
